requirements.txt
autopep8==1.5.7
dnspython==2.1.0
Flask==1.1.4
Flask-Cors==3.0.10
Flask-SQLAlchemy
flask-restplus==0.13.0
flask-migrate
Flask-Script==2.0.6
PyJWT==2.1.0
python-dotenv==0.17.1
requests==2.25.1
Werkzeug==0.16.1
psycopg2
